Remove hard-coded cluster fromt the Application model
I was unable to get Airflow databuilder loads working for my own tables, but was able to load them with the sample loader script.  I realized that the sample loader is always defaulting to the "gold" cluster, and so was databuilder, but my own tables were in a different cluster.  This removed the hard-coded cluster and makes it an option.

More discussion at https://amundsenworkspace.slack.com/archives/CHQERT0D7/p1588011296187400