Compatibility with latest Neo4J 4.0.4
Starting with Neo4J v4, it no longer auto generates SSL/TLS
certificates if they aren't supplied. To keep the example
working, I needed config to specify whether the connection
to Neo4J should be encrypted or not.

It defaults to False, at the moment, so I think most
people would need to add an explicit True to their config.