Set snowflake extractor database to be consistent with other extractors
This is a proposed fix for the bug described in https://github.com/lyft/amundsen/issues/494 - it adds a new configuration key, SNOWFLAKE_DATABASE_KEY, and uses it to set the database that metadata should be extracted from.  The DATABASE_KEY reverts back to simply describing the database, with a default of 'snowflake'.