Spike out PoC for Tableau integration
This patch adds a very barebones Tableau dashboard integration.

Currently, it supports adding workbooks, users and owners (which are fake), and
finding the last updated timestamp. To run it, startup Amundsen as normal and then
run `python example/scripts/sample_data_loader.py`. You'll need to make sure that
you've added your Tableau site config and credentials to the top of the sample loader
script. It also does not support creation of tables and columns at this time, so you'll
need to add those manually using a CSV job.