Add Tableau dashboard query extractor
This patch introduces a new Tableau dashboard extractor
for gathering the queries used in Tableau workbooks.

In Tableau, these are classified as "custom SQL tables", and appear as a
datasource for the sheets in which they are used. Luckily, the Tableau
Metadata API is smart enough to include the tables used in a custom SQL
query in the workbook <> table relation.

Signed-off-by: Carter Landis <carter.landis@gusto.com>