tests: HACK to preserve trailing whitespace in query test
I edit with trim_trailing_white_space_on_save enabled in Sublime, editing
anything else in this file will cause this sensitive test to break.

This isn't the prettiest way to fix this, but I think this is enough of a
gotcha that the previous status quo is not acceptable.

Very open to feedback if there's a better way or established pattern to fix
this footgun.

Signed-off-by: Dorian Johnson <2020@dorianj.net>