refactor: A factory for SQLAlchemyExtractors wrapped by other Extractors.
A common pattern has emerged where database-specific extractors delegate the execution
of queries to a SQLAlchemyExtractor instance. The DB specific extractor has the code
tailored to a DB's metadata table structure. The SQLAlchemyExtractor has the basic
connection lifecycle code.

The pattern involved a few lines of tricky-looking code to handle configs that looks
like it has been copy-pasted from one config to the next. This pulls that out in to
a factory function in the sql_alchemy_extractor module.

Signed-off-by: Dave Cameron <dcameron@digitalocean.com>