fix: Avoid error, check before calling close.
Without checking the existence of the _alchemy_extractor attribute, and that
the value is not None, an AttributeError can be raised when close() if called
if the init method didn't complete. Tasks still call close to clean up, so
this code should check if there is anything to clean up first.

Signed-off-by: Dave Cameron <dcameron@digitalocean.com>